+++ client/src/components/SearchResultPage/Views/NoResults.js
