+++ client/src/components/SearchResultPage/Views/SearchPage.css
