+++ client/src/components/SearchResultPage/Views/SearchPageView.js
