+++ client/src/components/SearchResultPage/Views/SearchResult.css
