+++ client/src/components/SearchResultPage/Views/SearchResultBukketlist.js
