+++ client/src/components/SearchResultPage/Views/SearchResultUser.js
